Added: Dynamic page load on homepage
