+++ content/posts/license.mdx
@@ -5,6 +5,8 @@ date: 2020-06-08
 author: herop-lab
 tags:
   - license
+thumbnail: https://main--cheerful-treacle-913a24.netlify.app/_next/static/media/news1.c585a41f.png
+excerpt: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
 ---
 
 ## Next.js and netlify blog template is managed under the MIT license
