+++ content/posts/markdown-syntax.mdx
@@ -6,6 +6,8 @@ author: herop-lab
 tags:
   - style
   - markdown
+thumbnail: https://main--cheerful-treacle-913a24.netlify.app/_next/static/media/advisory-board-post-image.44940d20.png
+excerpt: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
 ---
 
 This article offers a sample of basic Markdown syntax that can be used in Hugo content files, also it shows whether basic HTML elements are decorated with CSS in MDX.
