+++ content/posts/welcome.mdx
@@ -5,6 +5,8 @@ date: 2020-06-10
 author: herop-lab
 tags:
   - document
+thumbnail: https://main--cheerful-treacle-913a24.netlify.app/_next/static/media/fellowship-post-image.7e72dbb7.png
+excerpt: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
 ---
 
 ## The standard Lorem Ipsum passage
