+++ package.json
@@ -97,5 +97,10 @@
     "typescript": "^5.2.2",
     "yaml-loader": "^0.8.0"
   },
-  "prepare": "husky install"
+  "prepare": "husky install",
+  "browser": {
+    "fs": false,
+    "path": false,
+    "os": false
+  }
 }
