--- src/components/homepage/index.ts
@@ -1,3 +1,0 @@
-import HomePage from "./homepage";
-
-export default HomePage;
